backend/requirements.txt
fastapi>=0.115.8
uvicorn[standard]>=0.35.0
pandas>=2.2.3
pdfplumber>=0.11.5
tabula-py>=2.10.0
python-multipart>=0.0.20
pydantic[email]>=2.11.7
openai>=2.26.0
google-generativeai>=0.8.4
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
